Fix BitParameterBlock default value (uninitialized >32 bit)
64 bit BitParameterBlock were initializied with a 32 bit buffer leading
to a buffer overflow.

Use a 64 bit buffer instead.

+++ parameter/BitParameterBlock.cpp
@@ -60,8 +60,11 @@ size_t CBitParameterBlock::getSize() const
 // Used for simulation and virtual subsystems
 void CBitParameterBlock::setDefaultValues(CParameterAccessContext& parameterAccessContext) const
 {
-    // Get default value from type
-    uint32_t uiDefaultValue = 0;
+    // Default value is 0 as their is no min bound for bit parameters,
+    // thus 0 is always a valid value.
+    // BitParameterBlock can be as long a 64 bit, thus an 64 bit long variable
+    // is necessary to initialize it.
+    uint64_t uiDefaultValue = 0;
 
     // Write blackboard
     CParameterBlackboard* pBlackboard = parameterAccessContext.getParameterBlackboard();
